pointer targets in passing argument 6 of ‘recvfrom’ differ in signedness  well.. obviously on linux, socklen_t is size_t, ie. "unsigend int" and  not int
git-svn-id: http://svn.unix-ag.uni-kl.de/vpnc/trunk@108 315857ad-0bdb-0310-b42e-dec37551a5f0

+++ tunip.c
@@ -151,7 +151,7 @@ struct encap_method {
 	unsigned int bufsize, bufpayload, var_header_size;
 	int buflen;
 	struct sockaddr_in from;
-	int fromlen;
+	size_t fromlen;
 
 	int (*recv) (struct encap_method * encap,
 		unsigned char *buf, unsigned int bufsize, struct sockaddr_in * from);
@@ -429,7 +429,7 @@ static int encap_udp_new(struct encap_method *encap, int udp_fd)
  */
 int find_local_addr(struct sockaddr_in *dest, struct sockaddr_in *source)
 {
-	int addrlen;
+	size_t addrlen;
 	struct sockaddr_in dest_socket;
 	int fd;
